+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- `DateHandler` utils as it belongs to another package and will be replaced in the future
+- :boom: `DateHandler` utils as it belongs to another package and will be replaced in the future
 
 ## [2.10.0] - 2023-10-20